refactor: NameType -> TableName으로 네이밍 변경
